backend/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
python-dotenv==1.0.0
supabase==2.3.5
postgrest-py==0.12.0
PyJWT==2.8.1
requests==2.31.0
